Server: moved updateCredential helper into authHelpers, then implemented it into authRouter. Implemented hashing in the new /auth/update route and passed the hash into the helper as password to be stored for the update.
