+++ server/Controller/Routes/user.js
@@ -2,7 +2,7 @@ const userRouter = require('express').Router();
 
 // userRouter-Helpers
 const getUser = require('../../Helpers/userHelpers/getUser');
-const updateCredentials = require('../../Helpers/userHelpers/updateCredentials');
+const updateCredentials = require('../../Helpers/authHelpers/updateCredentials');
 const updateProfile = require('../../Helpers/userHelpers/updateProfile');
 
 // Get User
@@ -10,12 +10,6 @@ userRouter.get('/', (req,res) => {
     console.log("/user/ hit");
     const {_id} = req.body;
     getUser(_id, res);
-})
-
-// Update User Credentials
-userRouter.put('/', (req,res) => {
-    const {name, password} = req.body;
-    updateCredentials(name,password,res);
 })
 
 // Update User Profile
